bennaaym/labgraph_monitor: directories refactored, README updated, typos fixed
