+++ extensions/prototypes/labgraph_monitor/README.md
@@ -7,35 +7,42 @@ This extension is an interactive visualization tool to monitor and make real-tim
 **Prerequisites**:
 
 -   [Node.js](https://nodejs.org/en/)
--   [Yarn](https://classic.yarnpkg.com/lang/en/docs/install) (**RECOMMENDED**)
+-   [Yarn](https://classic.yarnpkg.com/lang/en/docs/install)
+
+Check that node and yarn were properly installed by running the following commands
+
+```
+node -v
+```
+
+```
+yarn -v
+```
 
 **Set up the application**
-
-(!) The following tutorial utilizes **yarn** (recommended) however **npm** can be utilized as well.
 
 1. Be sure that you are inside **extensions/prototypes/labgraph_monitor** directory
 
 ```
-labgraph> cd extensions/prototypes/labgraph_monitor
-labgraph\extensions\prototypes\labgraph_monitor>
+cd extensions/prototypes/labgraph_monitor
 ```
 
 2. Install dependencies by running **yarn** command
 
 ```
-labgraph\extensions\prototypes\labgraph_monitor> yarn
+yarn
 ```
 
 3. Test the application by running the following command
 
 ```
-labgraph\extensions\prototypes\labgraph_monitor> yarn test --watchAll=false
+yarn test --watchAll=false
 ```
 
 4. Run the application
 
 ```
-labgraph\extensions\prototypes\labgraph_monitor> yarn start
+yarn start
 ```
 
 (!) The application will be running on **localhost:3000** by default
@@ -44,7 +51,7 @@ labgraph\extensions\prototypes\labgraph_monitor> yarn start
 
 <image src="https://i.ibb.co/nBn4mv9/main-screen-frame.png" alt=""/>
 
-### Mocks :
+### Mocks:
 
 Mocks are a quick way to get familiar with the user interface. They do not require any connection to the LabGraph Websockets API and provide visualization for most of the existing [LabGraph examples](https://github.com/facebookresearch/labgraph/tree/main/labgraph/examples).
 
@@ -64,7 +71,7 @@ REACT_APP_WS_API="ws://127.0.0.1:9000"
 
 (!) LabGraph Websocket server runs on localhost:9000 by default
 
-2. Run Labgraph Websockets server. The following tutorial show how to run LabGraph Websocket server properly : [tutorial](https://github.com/facebookresearch/labgraph/pull/58/files#diff-247005c77570899ce53f81a83b2a5fe6e7535616cc96564d67378fe7f73dac49)
+2. Run Labgraph Websockets server. The following tutorial shows how to run LabGraph Websocket server properly : [tutorial](https://github.com/facebookresearch/labgraph/pull/58/files#diff-247005c77570899ce53f81a83b2a5fe6e7535616cc96564d67378fe7f73dac49)
 
 3. Under LabGraph Monitor settings panel click on REALTIME option and click connect.
 
@@ -75,7 +82,7 @@ To see the information related to a specific node or edge, just click on it, the
 <table>
   <tr>
     <td>Node</td>
-     <td>edge</td>
+     <td>Edge</td>
   </tr>
   <tr>
     <td><img src="https://i.ibb.co/MnB045Z/node-frame.png"></td>
