+++ extensions/prototypes/labgraph_monitor/src/redux/reducers/graph/common/interfaces/IGraph.ts
