+++ extensions/prototypes/labgraph_monitor/src/redux/reducers/graph/common/interfaces/INode.ts
