+++ extensions/prototypes/labgraph_monitor/src/redux/reducers/graph/ws/enums/WS_STATE.ts
