+++ extensions/prototypes/labgraph_monitor/src/redux/reducers/graph/ws/interfaces/IWS.ts
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import IGraph from './IGraph';
+import IGraph from '../../common/interfaces/IGraph';
 import WS_STATE from '../enums/WS_STATE';
 
 interface IWS {
