--- extensions/prototypes/labgraph_monitor/src/redux/reducers/mock/interfaces/IMock.ts
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-import IGraph from './IGraph';
-
-interface IMock {
-    mockGraph: IGraph;
-}
-
-export default IMock;
