--- extensions/prototypes/labgraph_monitor/src/redux/reducers/ws/interfaces/INode.ts
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-interface INode {
-    [name: string]: {
-        upstreams: {
-            [upstream: string]: Array<{
-                name: string;
-                fields: {
-                    [fieldName: string]: string;
-                };
-            }>;
-        };
-    };
-}
-
-export default INode;
